Added Meal Review Page. Added new color to resources.
