+++ app/src/main/java/com/example/munchbox/ui/MealReviewScreen.kt
@@ -8,11 +8,15 @@ import androidx.compose.foundation.layout.fillMaxHeight
 import androidx.compose.foundation.layout.fillMaxWidth
 import androidx.compose.foundation.layout.height
 import androidx.compose.foundation.layout.padding
+import androidx.compose.foundation.rememberScrollState
+import androidx.compose.foundation.verticalScroll
 import androidx.compose.material3.Button
 import androidx.compose.material3.Divider
 import androidx.compose.material3.OutlinedButton
 import androidx.compose.material3.Text
 import androidx.compose.runtime.Composable
+import androidx.compose.runtime.collectAsState
+import androidx.compose.runtime.getValue
 import androidx.compose.ui.Alignment
 import androidx.compose.ui.Modifier
 import androidx.compose.ui.platform.LocalContext
@@ -20,8 +24,12 @@ import androidx.compose.ui.res.dimensionResource
 import androidx.compose.ui.res.stringResource
 import androidx.compose.ui.text.font.FontWeight
 import androidx.compose.ui.tooling.preview.Preview
+import androidx.compose.ui.unit.dp
+import androidx.lifecycle.viewmodel.compose.viewModel
 import com.example.munchbox.R
+import com.example.munchbox.data.DataSource
 import com.example.munchbox.data.OrderUiState
+import com.example.munchbox.ui.components.OrderSummaries
 
 @Composable
 fun MealReviewScreen(
@@ -30,25 +38,42 @@ fun MealReviewScreen(
     onNextButtonClicked: () -> Unit = {},
     modifier: Modifier = Modifier
 ) {
-    Row(
+    val scrollState = rememberScrollState()
+    Column(
         modifier = Modifier
-            .padding(dimensionResource(R.dimen.padding_medium))
-    ) {
-        Column(
-            verticalArrangement = Arrangement.spacedBy(dimensionResource(R.dimen.padding_small))
+            .verticalScroll(scrollState),
+    ){
+        OrderSummaries(orderUiState, modifier)
+        Row(
+            modifier = Modifier
+                .padding(dimensionResource(R.dimen.padding_medium))
         ) {
-            Button(
-                modifier = Modifier.fillMaxWidth(),
-                onClick = onNextButtonClicked
+            Column(
+                verticalArrangement = Arrangement.spacedBy(dimensionResource(R.dimen.padding_small))
             ) {
-                Text(stringResource(R.string.send))
-            }
-            OutlinedButton(
-                modifier = Modifier.fillMaxWidth(),
-                onClick = onCancelButtonClicked
-            ) {
-                Text(stringResource(R.string.cancel))
+                Button(
+                    modifier = Modifier.fillMaxWidth(),
+                    onClick = onNextButtonClicked
+                ) {
+                    Text(stringResource(R.string.checkout))
+                }
+                OutlinedButton(
+                    modifier = Modifier.fillMaxWidth(),
+                    onClick = onCancelButtonClicked
+                ) {
+                    Text(stringResource(R.string.cancel))
+                }
             }
         }
     }
+}
+
+@Composable
+@Preview
+fun PreviewMealReviewScreen() {
+    val viewModel:OrderViewModel = viewModel()
+    val uiState by viewModel.uiState.collectAsState()
+    viewModel.setMeals(meals = DataSource.allMeals)
+    viewModel.setPickupOptions(pickupOptions = DataSource.pickUpOptions)
+    MealReviewScreen(orderUiState = uiState)
 }
